% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{add_locations}
\alias{add_locations}
\title{Add locations to dataset}
\usage{
add_locations(dataset, locations)
}
\arguments{
\item{dataset}{Dataset}

\item{locations}{Charater, valid HDX locations}
}
\value{
A Dataset
}
\description{
Add locations to dataset
}
\details{
Add locations to dataset
}
