% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{add_organization}
\alias{add_organization}
\title{Add organization to dataset}
\usage{
add_organization(dataset, organization)
}
\arguments{
\item{dataset}{Dataset}

\item{organization}{Charater, valid HDX organization}
}
\value{
A Dataset
}
\description{
Add organization to dataset
}
\details{
Add locations to dataset
}
\examples{
\dontrun{
}
}
