% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{add_resource}
\alias{add_resource}
\title{Add resource to dataset}
\usage{
add_resource(dataset, resource, ignore_dataset_id = FALSE)
}
\arguments{
\item{dataset}{}
}
\value{
A Dataset
}
\description{
Add resource to dataset
}
\details{
Delete resource from dataset
}
\examples{
\dontrun{
 # Setting the config to use HDX default server
 delete_resource(dataset, 1) # first resource
}

}
