% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{add_tags}
\alias{add_tags}
\title{Add tags to dataset}
\usage{
add_tags(dataset, tags)
}
\arguments{
\item{dataset}{Dataset}

\item{tags}{Charater, tags}
}
\value{
A Dataset
}
\description{
Add tags to dataset
}
\details{
Add tags to dataset
}
