% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocabulary.R
\name{authorized_tags}
\alias{authorized_tags}
\title{List all authorized tags in HDX}
\usage{
authorized_tags(configuration = NULL)
}
\arguments{
\item{configuration}{Configuration}
}
\value{
A vector of character, the authorized tags name
}
\description{
List all authorized tags in HDX
}
\examples{
\dontrun{
autorized_tags()
}
}
