% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R, R/organization.R, R/resource.R,
%   R/user.R, R/utils.R
\name{browse.Dataset}
\alias{browse.Dataset}
\alias{browse.Organization}
\alias{browse.Resource}
\alias{browse.User}
\alias{browse}
\alias{browse.default}
\title{Browse a HDX object}
\usage{
\method{browse}{Dataset}(x, ...)

\method{browse}{Organization}(x, ...)

\method{browse}{Resource}(x, ...)

\method{browse}{User}(x, ...)

browse(x, ...)

\method{browse}{default}(x, ...)
}
\arguments{
\item{x}{an HDX object}

\item{...}{Extra parameters}
}
\value{
Character Tags of the dataset
}
\description{
Browse a HDX object
}
\examples{
\dontrun{
# Setting the config to use HDX default server
 set_rhdx_config()
 res <- search_dataset(rows = 3L)
 browse(res[[1]])
}
}
