% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration.R
\name{Configuration}
\alias{Configuration}
\title{HDX Configuration}
\description{
HDX Configuration allow to connect to an HDX server and setup project where you can interact with the HDX platform
}
\details{
\strong{Methods}
\describe{
\item{\code{create(hdx_site, hdx_key, read_only, hdx_key_file, hdx_config, hdx_config_yaml, hdx_config_json, project_config, project_config_yaml, project_config_json, configuration)}}{
Create a HDX Configuration
}
\item{\code{get_credentials()}}{
Setup a HDX Configuration object
}
\item{\code{set_read_only(read_only)}}{
Read a Configuration object
}\preformatted{\\item\{`set_hdx_key(hdx_key)`\}\{
  Provides an API key
\}

\\item\{`load_hdx_key(path)`\}\{
  Provides an API key from path
\}

\\item\{`get_hdx_key()`\}\{
  Get the HDX server site in use.
\}

\\item\{`set_hdx_key(hdx_site)`\}\{
  Set the HDX server site to use.
\}
}

\item{\code{get_hdx_key()}}{
Get the API key
}

\item{\code{get_hdx_key_url()}}{
Get the HDX server URL
}

\item{\code{remoteclient()}}{
Remote API client
}

\item{\code{call_remoteclient()}}{
Call remote API client
}

\item{\code{setup()}}{
Setup configuration i.e HDX server, API and read only status.
}

\item{\code{delete()}}{
Delete actual configuration.
}

\item{\code{general_statistics()}}{
Some statistics about HDX
}

}
}
\examples{
\dontrun{
set_rhdx_config(hdx_site = "prod")
get_rhd_config()
}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Configuration$new()}}
\item \href{#method-get_credentials}{\code{Configuration$get_credentials()}}
\item \href{#method-set_read_only}{\code{Configuration$set_read_only()}}
\item \href{#method-set_hdx_key}{\code{Configuration$set_hdx_key()}}
\item \href{#method-get_hdx_key}{\code{Configuration$get_hdx_key()}}
\item \href{#method-load_hdx_key}{\code{Configuration$load_hdx_key()}}
\item \href{#method-set_hdx_site}{\code{Configuration$set_hdx_site()}}
\item \href{#method-get_hdx_site}{\code{Configuration$get_hdx_site()}}
\item \href{#method-get_hdx_site_url}{\code{Configuration$get_hdx_site_url()}}
\item \href{#method-remoteclient}{\code{Configuration$remoteclient()}}
\item \href{#method-call_action}{\code{Configuration$call_action()}}
\item \href{#method-call_remoteclient}{\code{Configuration$call_remoteclient()}}
\item \href{#method-read}{\code{Configuration$read()}}
\item \href{#method-setup}{\code{Configuration$setup()}}
\item \href{#method-delete}{\code{Configuration$delete()}}
\item \href{#method-get_global_config}{\code{Configuration$get_global_config()}}
\item \href{#method-general_statistics}{\code{Configuration$general_statistics()}}
\item \href{#method-as_list}{\code{Configuration$as_list()}}
\item \href{#method-print}{\code{Configuration$print()}}
\item \href{#method-clone}{\code{Configuration$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Configuration$new(
  hdx_site = "prod",
  hdx_key = NULL,
  hdx_config = NULL,
  hdx_config_file = NULL,
  read_only = TRUE,
  user_agent = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_credentials"></a>}}
\subsection{Method \code{get_credentials()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Configuration$get_credentials()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_read_only"></a>}}
\subsection{Method \code{set_read_only()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Configuration$set_read_only(read_only = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_hdx_key"></a>}}
\subsection{Method \code{set_hdx_key()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Configuration$set_hdx_key(hdx_key)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_hdx_key"></a>}}
\subsection{Method \code{get_hdx_key()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Configuration$get_hdx_key()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-load_hdx_key"></a>}}
\subsection{Method \code{load_hdx_key()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Configuration$load_hdx_key(path)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_hdx_site"></a>}}
\subsection{Method \code{set_hdx_site()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Configuration$set_hdx_site(hdx_site = "prod")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_hdx_site"></a>}}
\subsection{Method \code{get_hdx_site()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Configuration$get_hdx_site()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_hdx_site_url"></a>}}
\subsection{Method \code{get_hdx_site_url()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Configuration$get_hdx_site_url()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-remoteclient"></a>}}
\subsection{Method \code{remoteclient()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Configuration$remoteclient()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-call_action"></a>}}
\subsection{Method \code{call_action()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Configuration$call_action(action, ..., verb = "get")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-call_remoteclient"></a>}}
\subsection{Method \code{call_remoteclient()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Configuration$call_remoteclient(
  action,
  data = NULL,
  verb = "get",
  encode = "json",
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read"></a>}}
\subsection{Method \code{read()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Configuration$read()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setup"></a>}}
\subsection{Method \code{setup()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Configuration$setup(
  hdx_site = "prod",
  hdx_key = NULL,
  read_only = TRUE,
  hdx_config = NULL,
  configuration = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete"></a>}}
\subsection{Method \code{delete()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Configuration$delete()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_global_config"></a>}}
\subsection{Method \code{get_global_config()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Configuration$get_global_config()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-general_statistics"></a>}}
\subsection{Method \code{general_statistics()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Configuration$general_statistics()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-as_list"></a>}}
\subsection{Method \code{as_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Configuration$as_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Configuration$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Configuration$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
