% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{count_datasets}
\alias{count_datasets}
\title{Gives the number of datasets available}
\usage{
count_datasets(configuration = NULL)
}
\arguments{
\item{configuration}{Configuration, an HDX config object}
}
\value{
The number of datasets
}
\description{
Gives the number of datasets available
}
\examples{
\dontrun{
 set_rhdx_config()
 count_datasets()
}
}
