% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{create_dataset}
\alias{create_dataset}
\title{Create dataset from list}
\usage{
create_dataset(initial_data)
}
\arguments{
\item{initial_data}{List, list of data}
}
\value{
Dataset the dataset
}
\description{
Create dataset from list with required fields
}
\examples{
\dontrun{
# Setting the config to use HDX default server
 set_rhdx_config()
 metadata <- list(name = "hum-dataset", date = "09/25/2018", title = "Humanitarian dataset")
 create_data(metdata)
 res
}
}
