% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{create_in_hdx.Dataset}
\alias{create_in_hdx.Dataset}
\title{Create dataset in HDX}
\usage{
\method{create_in_hdx}{Dataset}(dataset, upload_resources = TRUE,
  verbose = FALSE)
}
\arguments{
\item{dataset}{Dataset}
}
\value{
an HDX dataset
}
\description{
Create dataset in HDX
}
\examples{
\dontrun{
# Setting the config to use HDX default server
 set_rhdx_config()
 res <- read_dataset("mali-3wop")
 res
} 
}
