% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organization.R
\name{create_in_hdx.Organization}
\alias{create_in_hdx.Organization}
\title{Create organization in HDX}
\usage{
\method{create_in_hdx}{Organization}(organization, verbose = FALSE)
}
\arguments{
\item{organization}{Organization}
}
\value{
an HDX organization
}
\description{
Create organization in HDX
}
\examples{
\dontrun{
} 
}
