% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{create_in_hdx.Resource}
\alias{create_in_hdx.Resource}
\title{Create resource in HDX}
\usage{
\method{create_in_hdx}{Resource}(resource, verbose = FALSE)
}
\arguments{
\item{resource}{Resource}
}
\value{
an HDX resource
}
\description{
Create resource in HDX
}
\examples{
\dontrun{
} 
}
