% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{create_resource}
\alias{create_resource}
\title{Create resource from list}
\usage{
create_resource(initial_data)
}
\arguments{
\item{initial_data}{List, list of data}
}
\value{
Resource the resource
}
\description{
Create resource from list
}
\examples{
\dontrun{
}
}
