% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration.R
\name{create_rhdx_config}
\alias{create_rhdx_config}
\title{Create an HDX configuration object}
\usage{
create_rhdx_config(hdx_site = "prod", hdx_key = NULL,
  read_only = TRUE, hdx_config = NULL, hdx_config_file = NULL)
}
\arguments{
\item{hdx_site}{Character to specify which HDX server you want to use. Default to "prod".}

\item{hdx_key}{Character for the CKAN API key, it is required to push data into HDX}

\item{read_only}{Logical if \code{FALSE} and hdx_key provided is correct you can push metdata and data to HDX}

\item{hdx_config}{List of HDX configuration}

\item{hdx_config_file}{Character, path of the HDX config file in JSON and YAML format}
}
\value{
An HDX Configuration object
}
\description{
Create and HDX configuration object
}
