% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{delete_all_resources}
\alias{delete_all_resources}
\title{Delete all resource from dataset}
\usage{
delete_all_resources(dataset)
}
\arguments{
\item{dataset}{}
}
\value{
Dataset without resources
}
\description{
Delete all resource from dataset
}
\details{
Delete all resources from dataset
}
\examples{
\dontrun{
 # Setting the config to use HDX default server
 delete_all_resources(dataset) # remove first resource
}

}
