% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{delete_from_hdx}
\alias{delete_from_hdx}
\title{Delete dataset from HDX}
\usage{
delete_from_hdx(dataset)
}
\arguments{
\item{dataset}{Dataset}
}
\value{
"None"
}
\description{
Delete dataset from HDX
}
\examples{
\dontrun{
 #Setting the config to use HDX default server
 set_rhdx_config(read_only = FALSE, hdx_key = "xxxx")
 res <- search_dataset(rows = 3L)
 delete_dataset(res[[1]])
}
}
