% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{delete_resource}
\alias{delete_resource}
\title{Delete resource from dataset}
\usage{
delete_resource(dataset, index)
}
\arguments{
\item{dataset}{}

\item{index}{Integer the index of the resource to be removed}

\item{dataset}{Dataset the dataset from which we one to remove the resource}
}
\value{
Dataset the dataset without the resource
}
\description{
Delete resource from dataset
}
\details{
Delete resource from dataset
}
\examples{
\dontrun{
 # Setting the config to use HDX default server
 delete_resource(dataset, 1) # remove first resource
}
}
