% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{delete_resources}
\alias{delete_resources}
\title{Delete all resource from dataset}
\usage{
delete_resources(dataset)
}
\arguments{
\item{dataset}{A Dataset, the dataset to remove}
}
\value{
Dataset without resources
}
\description{
Delete all resource from dataset
}
\details{
Delete all resources from dataset
}
