% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration.R
\name{delete_rhdx_config}
\alias{delete_rhdx_config}
\title{Delete rhdx config}
\usage{
delete_rhdx_config()
}
\value{
None
}
\description{
Delete the configuration settings for using rhdx.
}
\details{
Delete HDX config
}
\examples{
\dontrun{
# Setting the config to use HDX default server
set_rhdx_config(hdx_site = "prod")
get_rhdx_config()

delete_rhdx_config()
get_rhdx_config()
}
}
