% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{get_dataset_date}
\alias{get_dataset_date}
\title{Get the Dataset date}
\usage{
get_dataset_date(dataset)
}
\arguments{
\item{dataset}{Dataset}
}
\value{
Date, date of the dataset
}
\description{
Date of dataset
}
\examples{
\dontrun{
# Setting the config to use HDX default server
 set_rhdx_config()
 res <- search_dataset(rows = 3L)
 get_dataset_date(res[[1]])
}
}
