% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{get_locations_name}
\alias{get_locations_name}
\title{Dataset locations name}
\usage{
get_locations_name(dataset)
}
\arguments{
\item{dataset}{Dataset}
}
\value{
Character locations of the dataset
}
\description{
Gets locations name from the datasets
}
\examples{
\dontrun{
# Setting the config to use HDX default server
 set_rhdx_config()
 res <- search_dataset(rows = 3L)
 get_locations_name(res[[1]])
}
}
