% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{get_resource_dataset}
\alias{get_resource_dataset}
\title{Get the dataset containing the resource}
\usage{
get_resource_dataset(resource)
}
\arguments{
\item{resource}{Resource, an HDX resource}
}
\value{
a Dataset, the dataset containing the resource
}
\description{
Get the dataset containing the resource
}
