% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{get_resource_format}
\alias{get_resource_format}
\title{Get the file format of the resource}
\usage{
get_resource_format(resource)
}
\arguments{
\item{resource}{Resource, an HDX resource}
}
\value{
A character, the format of the resource
}
\description{
Get the file format of the resource
}
