% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{get_resource_layers}
\alias{get_resource_layers}
\title{List layers available in spatial resources on HDX}
\usage{
get_resource_layers(resource, download_folder = NULL, quiet = TRUE)
}
\arguments{
\item{resource}{Resource, an HDX resource}

\item{download_folder}{Character, path of the directory where you will store the data}

\item{quiet}{Logical, no progress bar from download (default = FALSE)}
}
\value{
the layers name
}
\description{
List layers available in spatial resources on HDX
}
