% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{get_resource_sheets}
\alias{get_resource_sheets}
\title{Get the names of the sheets of XLS(X) resources}
\usage{
get_resource_sheets(resource, folder = NULL, quiet = TRUE, ...)
}
\arguments{
\item{resource}{Resource, an HDX resource}

\item{folder}{Character, path of the directory where you will store the data}

\item{quiet}{Logical, no progress bar from download (default = FALSE)}

\item{...}{extra parameters}
}
\value{
the names of the sheets of XLS(X) resources
}
\description{
Get the names of the sheets of XLS(X) resources
}
