% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{get_resources}
\alias{get_resources}
\title{Add resource to dataset}
\usage{
get_resources(dataset, pattern = NULL, format = NULL)
}
\arguments{
\item{dataset}{Dataset}

\item{pattern}{regex pattern in resource name}

\item{format}{format of the resources}
}
\value{
resource_list
}
\description{
Add resource to dataset
}
