% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{get_tags_name}
\alias{get_tags_name}
\title{Dataset tags name}
\usage{
get_tags_name(dataset)
}
\arguments{
\item{dataset}{Dataset}
}
\value{
Character Tags of the dataset
}
\description{
Gets dataset tags name
}
\examples{
\dontrun{
# Setting the config to use HDX default server
 set_rhdx_config()
 res <- search_dataset(rows = 3L)
 get_tags_name(res[[1]])
}
}
