% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration.R
\name{HDXConfig}
\alias{HDXConfig}
\title{HDX Configuration}
\description{
HDX Configuration

HDX Configuration
}
\details{
HDX Configuration allow to connect to an HDX server
and setup project where you can interact with the HDX platform
}
\examples{
\dontrun{
set_rhdx_config(hdx_site = "prod")
get_rhd_config()
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{all info in list.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{HDXConfig$new()}}
\item \href{#method-get_credentials}{\code{HDXConfig$get_credentials()}}
\item \href{#method-set_read_only}{\code{HDXConfig$set_read_only()}}
\item \href{#method-set_hdx_key}{\code{HDXConfig$set_hdx_key()}}
\item \href{#method-get_hdx_key}{\code{HDXConfig$get_hdx_key()}}
\item \href{#method-set_hdx_site}{\code{HDXConfig$set_hdx_site()}}
\item \href{#method-get_hdx_site}{\code{HDXConfig$get_hdx_site()}}
\item \href{#method-get_hdx_site_url}{\code{HDXConfig$get_hdx_site_url()}}
\item \href{#method-remoteclient}{\code{HDXConfig$remoteclient()}}
\item \href{#method-call_action}{\code{HDXConfig$call_action()}}
\item \href{#method-read}{\code{HDXConfig$read()}}
\item \href{#method-setup}{\code{HDXConfig$setup()}}
\item \href{#method-delete}{\code{HDXConfig$delete()}}
\item \href{#method-get_global_config}{\code{HDXConfig$get_global_config()}}
\item \href{#method-general_statistics}{\code{HDXConfig$general_statistics()}}
\item \href{#method-as_list}{\code{HDXConfig$as_list()}}
\item \href{#method-print}{\code{HDXConfig$print()}}
\item \href{#method-clone}{\code{HDXConfig$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new Configuration object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HDXConfig$new(
  hdx_site = "prod",
  hdx_key = NULL,
  hdx_config = NULL,
  hdx_config_file = NULL,
  read_only = TRUE,
  user_agent = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{hdx_site}}{character the server instance to use}

\item{\code{hdx_key}}{character, the HDX API key}

\item{\code{hdx_config}}{configuration in a list}

\item{\code{hdx_config_file}}{a character value config file.
default is the config supplied in the package}

\item{\code{read_only}}{a logical value indicating if you want to just read
or be also able to write on the HDX server. You will need a API key to write.}

\item{\code{user_agent}}{a character value, User agent}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new Configuration object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_credentials"></a>}}
\if{latex}{\out{\hypertarget{method-get_credentials}{}}}
\subsection{Method \code{get_credentials()}}{
Configuration credentials when using a HDX API key
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HDXConfig$get_credentials()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
the username and password associated to the HDX API key
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_read_only"></a>}}
\if{latex}{\out{\hypertarget{method-set_read_only}{}}}
\subsection{Method \code{set_read_only()}}{
Create or revoke read only status
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HDXConfig$set_read_only(read_only = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{read_only}}{a logical value indicating if you want to just read or be also able to write on the HDX server. You will need a API key to write.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_hdx_key"></a>}}
\if{latex}{\out{\hypertarget{method-set_hdx_key}{}}}
\subsection{Method \code{set_hdx_key()}}{
Specify a HDX API key
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HDXConfig$set_hdx_key(hdx_key)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{hdx_key}}{a character with key}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_hdx_key"></a>}}
\if{latex}{\out{\hypertarget{method-get_hdx_key}{}}}
\subsection{Method \code{get_hdx_key()}}{
Specify a HDX API key
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HDXConfig$get_hdx_key()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a character, the HDX API key
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_hdx_site"></a>}}
\if{latex}{\out{\hypertarget{method-set_hdx_site}{}}}
\subsection{Method \code{set_hdx_site()}}{
Specify a HDX server to use
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HDXConfig$set_hdx_site(hdx_site = "prod")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{hdx_site}}{a character, the server type to use,
\code{prod}, \code{test}, \code{feature} or \code{demo}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a character, the HDX API key
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_hdx_site"></a>}}
\if{latex}{\out{\hypertarget{method-get_hdx_site}{}}}
\subsection{Method \code{get_hdx_site()}}{
Get the HDX server in use
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HDXConfig$get_hdx_site()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
the server type
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_hdx_site_url"></a>}}
\if{latex}{\out{\hypertarget{method-get_hdx_site_url}{}}}
\subsection{Method \code{get_hdx_site_url()}}{
Get the HDX server URL in use
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HDXConfig$get_hdx_site_url()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
the server URL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-remoteclient"></a>}}
\if{latex}{\out{\hypertarget{method-remoteclient}{}}}
\subsection{Method \code{remoteclient()}}{
Get the remoteclient currently used
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HDXConfig$remoteclient()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a crul::HttpClient
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-call_action"></a>}}
\if{latex}{\out{\hypertarget{method-call_action}{}}}
\subsection{Method \code{call_action()}}{
Call the client to the HDX API
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HDXConfig$call_action(action, ..., verb = "get")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{action}}{a character}

\item{\code{...}}{parameters for each verb used}

\item{\code{verb}}{a character the verb used, \code{post}, \code{get}, \code{put} or \code{patch}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
list a with status code and results
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read"></a>}}
\if{latex}{\out{\hypertarget{method-read}{}}}
\subsection{Method \code{read()}}{
read and show Configuration object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HDXConfig$read()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Configuration object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setup"></a>}}
\if{latex}{\out{\hypertarget{method-setup}{}}}
\subsection{Method \code{setup()}}{
Setup Configuration object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HDXConfig$setup(
  hdx_site = "prod",
  hdx_key = NULL,
  read_only = TRUE,
  hdx_config = NULL,
  configuration = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{hdx_site}}{a character value, the server}

\item{\code{hdx_key}}{a character value, the API key}

\item{\code{read_only}}{a logical value read only}

\item{\code{hdx_config}}{a list}

\item{\code{configuration}}{a character}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete"></a>}}
\if{latex}{\out{\hypertarget{method-delete}{}}}
\subsection{Method \code{delete()}}{
Delete a Configuration object
Access the global Configuration
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HDXConfig$delete()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_global_config"></a>}}
\if{latex}{\out{\hypertarget{method-get_global_config}{}}}
\subsection{Method \code{get_global_config()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HDXConfig$get_global_config()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list with HDX configuration information
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-general_statistics"></a>}}
\if{latex}{\out{\hypertarget{method-general_statistics}{}}}
\subsection{Method \code{general_statistics()}}{
Get general  statistics about the server
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HDXConfig$general_statistics()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list with statistics about the server
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-as_list"></a>}}
\if{latex}{\out{\hypertarget{method-as_list}{}}}
\subsection{Method \code{as_list()}}{
Convert configuration to list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HDXConfig$as_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
configuration in list format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print Configuration object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HDXConfig$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HDXConfig$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
