% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdxobject.R
\name{HDXObject}
\alias{HDXObject}
\title{HDXObject abstract class}
\description{
HDXObject abstract class

HDXObject abstract class
}
\details{
HDXObject class containing all logic for accessing,
creating, and updating HDX objects.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{placeholder for HDXObject field element}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{HDXObject$new()}}
\item \href{#method-update_from_yaml}{\code{HDXObject$update_from_yaml()}}
\item \href{#method-update_from_json}{\code{HDXObject$update_from_json()}}
\item \href{#method-get_required_fields}{\code{HDXObject$get_required_fields()}}
\item \href{#method-check_required_field}{\code{HDXObject$check_required_field()}}
\item \href{#method-as_list}{\code{HDXObject$as_list()}}
\item \href{#method-browse}{\code{HDXObject$browse()}}
\item \href{#method-get_configuration}{\code{HDXObject$get_configuration()}}
\item \href{#method-print}{\code{HDXObject$print()}}
\item \href{#method-clone}{\code{HDXObject$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new HDXObject object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HDXObject$new(initial_data = NULL, configuration = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{initial_data}}{list with required field to create a HDXObject}

\item{\code{configuration}}{a Configuration object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A HDXObject object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_from_yaml"></a>}}
\if{latex}{\out{\hypertarget{method-update_from_yaml}{}}}
\subsection{Method \code{update_from_yaml()}}{
Update metadata from yaml file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HDXObject$update_from_yaml(path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{(character) Path to YAML metadata}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_from_json"></a>}}
\if{latex}{\out{\hypertarget{method-update_from_json}{}}}
\subsection{Method \code{update_from_json()}}{
Update metadata from json file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HDXObject$update_from_json(path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{(character) Path to JSON metadata}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_required_fields"></a>}}
\if{latex}{\out{\hypertarget{method-get_required_fields}{}}}
\subsection{Method \code{get_required_fields()}}{
Get HDXObject required fields
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HDXObject$get_required_fields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list of required fields for a resource
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-check_required_field"></a>}}
\if{latex}{\out{\hypertarget{method-check_required_field}{}}}
\subsection{Method \code{check_required_field()}}{
Check HDXobject required field
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HDXObject$check_required_field()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a logical value, TRUE if the the resource is not missing
a required field and throws an error otherwise
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-as_list"></a>}}
\if{latex}{\out{\hypertarget{method-as_list}{}}}
\subsection{Method \code{as_list()}}{
Get HDXObject field into list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HDXObject$as_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a list with HDXObject field
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-browse"></a>}}
\if{latex}{\out{\hypertarget{method-browse}{}}}
\subsection{Method \code{browse()}}{
Browse HDX
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HDXObject$browse()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_configuration"></a>}}
\if{latex}{\out{\hypertarget{method-get_configuration}{}}}
\subsection{Method \code{get_configuration()}}{
Get the current configuration in use
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HDXObject$get_configuration()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A configuration object, the configuration in use
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print a Dataset object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HDXObject$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HDXObject$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
