% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showcase.R
\name{HDXShowcase}
\alias{HDXShowcase}
\title{HDX Showcase}
\description{
HDX Showcase

HDX Showcase
}
\details{
HDX Showcase
}
\section{Super class}{
\code{\link[rhdx:HDXObject]{rhdx::HDXObject}} -> \code{HDXShowcase}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{datasets}}{list of datasets using this showcase}

\item{\code{data}}{the field info into list}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{HDXShowcase$new()}}
\item \href{#method-get_datasets}{\code{HDXShowcase$get_datasets()}}
\item \href{#method-browse}{\code{HDXShowcase$browse()}}
\item \href{#method-as_list}{\code{HDXShowcase$as_list()}}
\item \href{#method-print}{\code{HDXShowcase$print()}}
\item \href{#method-clone}{\code{HDXShowcase$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="rhdx" data-topic="HDXObject" data-id="check_required_field">}\href{../../rhdx/html/HDXObject.html#method-check_required_field}{\code{rhdx::HDXObject$check_required_field()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rhdx" data-topic="HDXObject" data-id="get_configuration">}\href{../../rhdx/html/HDXObject.html#method-get_configuration}{\code{rhdx::HDXObject$get_configuration()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rhdx" data-topic="HDXObject" data-id="get_required_fields">}\href{../../rhdx/html/HDXObject.html#method-get_required_fields}{\code{rhdx::HDXObject$get_required_fields()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rhdx" data-topic="HDXObject" data-id="update_from_json">}\href{../../rhdx/html/HDXObject.html#method-update_from_json}{\code{rhdx::HDXObject$update_from_json()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rhdx" data-topic="HDXObject" data-id="update_from_yaml">}\href{../../rhdx/html/HDXObject.html#method-update_from_yaml}{\code{rhdx::HDXObject$update_from_yaml()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new Showcase object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HDXShowcase$new(initial_data = NULL, configuration = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{initial_data}}{list, data with required field to create Showcase}

\item{\code{configuration}}{Configuration, configuration to use}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a new Showcase object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_datasets"></a>}}
\if{latex}{\out{\hypertarget{method-get_datasets}{}}}
\subsection{Method \code{get_datasets()}}{
List datasets using the Showcase
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HDXShowcase$get_datasets()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a list of dataset
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-browse"></a>}}
\if{latex}{\out{\hypertarget{method-browse}{}}}
\subsection{Method \code{browse()}}{
Browse the Showcase page on HDX
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HDXShowcase$browse()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-as_list"></a>}}
\if{latex}{\out{\hypertarget{method-as_list}{}}}
\subsection{Method \code{as_list()}}{
Get dataset field into list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HDXShowcase$as_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a list with showcase field info
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print a Showcase object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HDXShowcase$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HDXShowcase$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
