% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{list_datasets}
\alias{list_datasets}
\title{List datasets}
\usage{
list_datasets(limit = NULL, offset = NULL, configuration = NULL)
}
\arguments{
\item{configuration}{an HDX configuration object}

\item{identifier}{character dataset keyword}
}
\value{
Dataset the dataset
}
\description{
List datasets
}
\examples{
\dontrun{
# Setting the config to use HDX default server
 set_rhdx_config()
 res <- read_dataset("mali-3wop")
 res
} 
}
