% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location.R
\name{list_locations}
\alias{list_locations}
\title{List locations}
\usage{
list_locations(
  sort = "name asc",
  all_fields = FALSE,
  configuration = NULL,
  ...
)
}
\arguments{
\item{sort}{Character sorting of the search results. Default: “name asc”, the allowed fields are ‘name’, ‘package_count’ and ‘title’}

\item{all_fields}{Logical if TRUE returns list instead of just names}

\item{configuration}{a Configuration}

\item{...}{Extra parameters to group_list
https://docs.ckan.org/en/ckan-2.8.2/api/index.html#ckan.logic.action.get.group_list}
}
\value{
A vector of locations names
}
\description{
List locations
}
\examples{
\dontrun{
# Setting the config to use HDX default server
 set_rhdx_config()
 list_locations(limit = 10L)
}
}
