% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showcase.R
\name{.list_showcases}
\alias{.list_showcases}
\alias{list_showcases}
\title{List showcases}
\usage{
.list_showcases(configuration = NULL)

list_showcases(configuration = NULL)
}
\arguments{
\item{configuration}{a Configuration}

\item{limit}{Integer limit}

\item{offset}{Integer offset}
}
\value{
A vector of showcases names
}
\description{
List showcases
}
\examples{
\dontrun{
# Setting the config to use HDX default server
 set_rhdx_config()
 list_showcases()
}
}
