% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocabulary.R
\name{list_vocabularies}
\alias{list_vocabularies}
\title{List available HDX Vocabulary objects}
\usage{
list_vocabularies(identifier = NULL, configuration = NULL)
}
\arguments{
\item{identifier}{character identifier}

\item{configuration}{Configuration}
}
\value{
A list of Vocabulary object
}
\description{
List available HDX Vocabulary objects
}
\examples{
\dontrun{
# Setting the config to use HDX default server
 set_rhdx_config()
 res <- list_vocabularies()
 res
}
}
