% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organization.R
\name{Organization}
\alias{Organization}
\title{HDX Organization}
\description{
HDX Organization
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Organization$new()}}
\item \href{#method-get_datasets}{\code{Organization$get_datasets()}}
\item \href{#method-browse}{\code{Organization$browse()}}
\item \href{#method-as_list}{\code{Organization$as_list()}}
\item \href{#method-print}{\code{Organization$print()}}
\item \href{#method-clone}{\code{Organization$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Organization$new(initial_data = NULL, configuration = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_datasets"></a>}}
\subsection{Method \code{get_datasets()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Organization$get_datasets()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-browse"></a>}}
\subsection{Method \code{browse()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Organization$browse()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-as_list"></a>}}
\subsection{Method \code{as_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Organization$as_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Organization$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Organization$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
