% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{pull_dataset}
\alias{pull_dataset}
\title{Read dataset}
\usage{
pull_dataset(identifier, configuration = NULL, ...)
}
\arguments{
\item{identifier}{character dataset keyword}

\item{configuration}{a Configuration object}
}
\value{
Dataset the dataset
}
\description{
Read an HDX dataset from its name or id
}
\examples{
\dontrun{
# Setting the config to use HDX default server
 set_rhdx_config()
 res <- read_dataset("mali-3wop")
 res
}
}
