% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organization.R
\name{pull_organization}
\alias{pull_organization}
\title{Read an HDX organization}
\usage{
pull_organization(identifier = NULL, include_datasets = FALSE,
  configuration = NULL, ...)
}
\arguments{
\item{identifier}{character resource uuid}

\item{include_datasets}{Logical, include datasets if TRUE}

\item{configuration}{an HDX configuration object}

\item{...}{Extra parameters}
}
\value{
HDX organization
}
\description{
Read an HDX organization
}
