% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{pull_resource}
\alias{pull_resource}
\title{Read an HDX resource}
\usage{
pull_resource(identifier, configuration = NULL)
}
\arguments{
\item{identifier}{Character resource uuid}

\item{configuration}{a Configuration object}
}
\value{
Resource
}
\description{
Read an HDX resource
}
\examples{
\dontrun{
#Setting the config to use HDX default server
 set_rhdx_config()
 res <- pull_resource("98aa1742-b5d3-40c3-94c6-01e31ded6e84")
 res
}
}
