% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showcase.R
\name{pull_showcase}
\alias{pull_showcase}
\title{Read Showcase}
\usage{
pull_showcase(identifier = NULL, configuration = NULL)
}
\arguments{
\item{identifier}{Character Showcase name or id}

\item{configuration}{Configuration an HDX configuration object}
}
\value{
A showcase
}
\description{
Read HDX Showcase
}
\details{
Delete resource from dataset
}
\examples{
\dontrun{
 # Setting the config to use HDX default server
 delete_resource(dataset, 1) # first resource
}
}
