% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocabulary.R
\name{pull_vocabulary}
\alias{pull_vocabulary}
\title{Read an HDX Vocabulary}
\usage{
pull_vocabulary(identifier = NULL, configuration = NULL)
}
\arguments{
\item{identifier}{character the name or id of the vocabulary}

\item{configuration}{Configuration}
}
\value{
The Vocabulary object
}
\description{
Read an HDX vocabulary from its name or id
}
\examples{
\dontrun{
# Setting the config to use HDX default server
 set_rhdx_config()
 res <- pull_vocabulary("xxxx")
 res
}
}
