% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{push_dataset}
\alias{push_dataset}
\title{Create dataset in HDX}
\usage{
push_dataset(dataset, upload_resources = TRUE, verbose = FALSE)
}
\arguments{
\item{dataset}{Dataset}
}
\value{
an HDX dataset
}
\description{
Create dataset in HDX
}
\examples{
\dontrun{
# Setting the config to use HDX default server
 set_rhdx_config(hdx_site = "demo", hdx_key = "xxxxxxxxxxxxxx", read_only = FALSE)
 res <- read_dataset("mali-3wop")
 push_dataset(res)
}
}
