% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organization.R
\name{push_organization}
\alias{push_organization}
\title{Create organization in HDX}
\usage{
push_organization(organization, verbose = FALSE)
}
\arguments{
\item{organization}{Organization}

\item{verbose}{Logical, if TRUE verbose output}
}
\value{
an HDX organization
}
\description{
Create organization in HDX
}
