% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{read_dataset}
\alias{read_dataset}
\title{Read dataset}
\usage{
read_dataset(identifier, configuration = NULL, ...)
}
\arguments{
\item{dataset}{Dataset}
}
\value{
Dataset the dataset
}
\description{
Read dataset
}
\examples{

\dontrun{
# Setting the config to use HDX default server
 set_rhdx_config()
 res <- read_dataset("mali-3wop")
 res
}

}
