% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_from_hdx}
\alias{read_from_hdx}
\title{Function to search HDX object}
\usage{
read_from_hdx(identifier, configuration = NULL, type = "dataset")
}
\description{
Function to search HDX object
}
\examples{
\dontrun{
 read_from_hdx("ocha-mali", type = "organization")
 }

}
