% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organization.R
\name{read_organization}
\alias{read_organization}
\title{Read an HDX resource}
\usage{
read_organization(identifier = NULL, include_datasets = FALSE,
  configuration = NULL, ...)
}
\arguments{
\item{identifier}{character resource uuid}

\item{configuration}{an HDX configuration object}
}
\value{
HDX organization
}
\description{
Read an HDX resource
}
\examples{
\dontrun{
#Setting the config to use HDX default server
 set_rhdx_config()
 res <- read_organization("wfp")
 res
}

}
