% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{read_resource}
\alias{read_resource}
\title{Read resource}
\usage{
read_resource(
  resource,
  sheet = NULL,
  layer = NULL,
  folder = NULL,
  simplify_json = TRUE,
  force_download = FALSE,
  ...
)
}
\arguments{
\item{resource}{Resource, an HDX resource}

\item{sheet}{Character, the name of the sheet to read if XLS(X) resources. The first sheet is read by default.}

\item{layer}{Character, the name of the layer to read if spatial data. The first sheet is read by default.}

\item{folder}{Character, the path of the folder to store the downloaded data}

\item{simplify_json}{Logical, if TRUE simplifies nested lists into vectors and data frames for JSON resources}

\item{force_download}{Logical, force download if TRUE}

\item{...}{Extra parameters}
}
\value{
an \code{tbl}, a \code{list} or a \code{sf} object depending on the type of resource you are reading
}
\description{
Read resource
}
