% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{read_user}
\alias{read_user}
\title{Read user}
\usage{
read_user(identifier, include_datasets = FALSE, configuration = NULL,
  ...)
}
\arguments{
\item{identifier}{character user keyword}

\item{configuration}{a Configuration object}
}
\value{
User the user
}
\description{
Read an HDX user from its name or id
}
\examples{
\dontrun{
# Setting the config to use HDX default server
 set_rhdx_config()
 res <- read_user("mali-3wop")
 res
}
}
