% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{refine_search}
\alias{refine_search}
\title{Filter a list of HDX datasets}
\usage{
refine_search(datasets_list, format = NULL, locations = NULL,
  hxl = FALSE, tags = NULL, organization = NULL, cod = FALSE,
  requestable = FALSE)
}
\arguments{
\item{datasets_list}{A list of dataset}

\item{format}{Character Format of a resource in the dataset}

\item{locations}{Character Locations of the dataset}

\item{hxl}{logical dataset with HXL tags}

\item{tags}{character Dataset with specified tags}

\item{organization}{Character Organizations sharing the dataset}

\item{cod}{logical dataset with COD tags}

\item{requestable}{logical requestable dataset}
}
\value{
Dataset the dataset
}
\description{
Filter a list of HDX datasets
}
\examples{
\dontrun{
# Setting the config to use HDX default server
 set_rhdx_config()
 res <- read_dataset("mali-3wop")
 res
}
}
