% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhdx-package.R
\docType{package}
\name{rhdx-package}
\alias{rhdx-package}
\alias{rhdx}
\title{rhdx}
\description{
R client for the Humanitarian Data Exchange platform
}
\author{
\email{dicko.ahmadou@gmail.com}
}
\keyword{package}
