% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching.R
\name{rhdx_cache}
\alias{rhdx_cache}
\alias{rhdx_cache_set_dir}
\alias{rhdx_cache_get_dir}
\alias{rhdx_cache_list}
\alias{rhdx_cache_delete}
\alias{rhdx_cache_clear}
\title{Caching HDX downloaded files}
\usage{
rhdx_cache_set_dir(path)

rhdx_cache_get_dir()

rhdx_cache_list()

rhdx_cache_delete(file)

rhdx_cache_clear()
}
\arguments{
\item{path}{Character directory to set}

\item{file}{Character, the file to delete}
}
\value{
the cache directory

the cache directory

list of files in the cache
}
\description{
Manage cached HDX downloaded files
}
\details{
The default cache directory is
\verb{~/.cache/R/rhdx_cache}, but you can set
your own path using \code{rhdx_cache_set_dir()}
}
\examples{
\dontrun{
rhdx_cache
## change the default cache directory
tmp <- tempdir()
rhdx_cache_set_dir(tmp)

## print current cache directory
rhdx_cache_get_dir()

## List available files in the current cache directory
rhdx_cache_list()

l <- rhdx_cache_list()[1] ## get the first file
rhdx_cache_delete(l) ## delete it

rhdx_cache_clear() ## delete all cached files
}
}
