% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{search_datasets}
\alias{search_datasets}
\title{Search for datasets on HDX}
\usage{
search_datasets(
  query = "*:*",
  filter_query = NULL,
  rows = 10L,
  page_size = 1000L,
  configuration = NULL,
  ...
)
}
\arguments{
\item{query}{Query terms, use solr format and default to "\emph{:}" (match everything)}

\item{filter_query}{Filter Query results}

\item{rows}{Number of matching records to return. Defaults to 10.}

\item{page_size}{Integer Size of page to return. Defaults to 1000.}

\item{configuration}{Configuration object.}

\item{...}{Extra parameters}
}
\value{
A list of HDX datasets
}
\description{
Search for datasets on HDX
}
\details{
Search and find datasets on HDX
}
\examples{
\dontrun{
 # Setting the config to use HDX default server
 search_datasets("displaced nigeria", rows = 3L)
}
}
