% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{search_in_hdx}
\alias{search_in_hdx}
\title{Function to search HDX object}
\usage{
search_in_hdx(query = "*:*", rows = 10L, page_size = 1000L,
  configuration = NULL, type = "dataset")
}
\description{
Function to search HDX object
}
\examples{
\dontrun{
search_in_hdx("mali idps", type = "dataset") \%>\%
 filter("ocha-mali" \%in\% organization, "idp" \%in\% tags) \%>\%
 get_resources() \%>\%
 filter("zipped shapefile" \%in\% format) \%>\%
 read() -> mali_idps_shp

 search_in_hdx("ocha mali", type = "organization")
 }

}
