% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{.search_resources}
\alias{.search_resources}
\alias{search_resources}
\title{Search resources}
\usage{
.search_resources(query = "*:*", configuration = NULL, ...)

search_resources(query = "*:*", configuration = NULL, ...)
}
\arguments{
\item{query}{Character, a query}

\item{configuration}{an HDX configuration object}

\item{...}{extra params}
}
\description{
Search Resources
}
