% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag.R
\name{search_tags}
\alias{search_tags}
\title{Search for datasets on HDX}
\usage{
search_tags(
  query = "",
  vocabulary_id = NULL,
  limit = NULL,
  offset = NULL,
  configuration = NULL
)
}
\arguments{
\item{query}{(character) - character to search for}

\item{vocabulary_id}{(character) - the id or name of the tag vocabulary to search in}

\item{limit}{(integer) - the maximum number of tags to return}

\item{offset}{(integer) - when \code{limit} is given, the offset to start returnings tags from}

\item{configuration}{Configuration object.}
}
\value{
A list of HDX tags
}
\description{
Search for datasets on HDX
}
\details{
Search and find tags on HDX
}
\examples{
\dontrun{
 # Setting the config to use HDX default server
 search_tags("idps", rows = 3L)
}
}
