% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration.R
\name{set_rhdx_config}
\alias{set_rhdx_config}
\title{Set rhdx config}
\usage{
set_rhdx_config(hdx_site = "prod", hdx_key = NULL, read_only = TRUE,
  hdx_config = NULL, hdx_config_file = NULL, configuration = NULL)
}
\arguments{
\item{hdx_site}{Character to specify which HDX server you want to use. Default to "prod".}

\item{hdx_key}{Character for the CKAN API key, it is required to push data into HDX}

\item{read_only}{Logical if `FALSE` and hdx_key provided is correct you can push metdata and data to HDX}

\item{hdx_config_file}{Character config file path, it can either a `JSON` or a `YAML`.}

\item{configuration}{Configuration object.}
}
\value{
Invisibly returns the rhdx config object
}
\description{
Sets the configuration settings for using rhdx.
}
\details{
Setting up a configuration will help you access and push data into HDX
}
\examples{

\dontrun{
# Setting the config to use HDX default server
set_rhdx_config(hdx_site = "prod")


# You can check your configuration using \\code{get_rhdx_config}
config <- get_rhdx_config()
}

}
