% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showcase.R
\docType{data}
\name{Showcase}
\alias{Showcase}
\title{Create and manipulate HDX Showcase}
\description{
Create and manipulate HDX Showcase
}
\details{
**Methods**
  \describe{
  }
}
\examples{
\dontrun{
}

}
\keyword{datasets}
