% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag.R
\name{Tag}
\alias{Tag}
\title{HDX Tag}
\description{
HDX Tag

HDX Tag
}
\details{
HDX tag
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{list of tag field element}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Tag$new()}}
\item \href{#method-as_list}{\code{Tag$as_list()}}
\item \href{#method-print}{\code{Tag$print()}}
\item \href{#method-clone}{\code{Tag$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new Tag object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tag$new(initial_data = NULL, configuration = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{initial_data}}{list data with required fields to create a tag object}

\item{\code{configuration}}{Configuration configuration to use}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a Tag object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-as_list"></a>}}
\subsection{Method \code{as_list()}}{
Tag object to list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tag$as_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a list with of tag fields element
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
Print a Tag object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tag$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{

}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tag$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
