% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocabulary.R
\name{Vocabulary}
\alias{Vocabulary}
\title{HDX Vocabulary}
\description{
HDX Vocabulary

HDX Vocabulary
}
\details{
HDX Vocabulary
}
\section{Super class}{
\code{\link[rhdx:HDXObject]{rhdx::HDXObject}} -> \code{Vocabulary}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{list of tag field element}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Vocabulary$new()}}
\item \href{#method-as_list}{\code{Vocabulary$as_list()}}
\item \href{#method-authorized_tags_name}{\code{Vocabulary$authorized_tags_name()}}
\item \href{#method-print}{\code{Vocabulary$print()}}
\item \href{#method-clone}{\code{Vocabulary$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="rhdx" data-topic="HDXObject" data-id="browse">}\href{../../rhdx/html/HDXObject.html#method-browse}{\code{rhdx::HDXObject$browse()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rhdx" data-topic="HDXObject" data-id="check_required_field">}\href{../../rhdx/html/HDXObject.html#method-check_required_field}{\code{rhdx::HDXObject$check_required_field()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rhdx" data-topic="HDXObject" data-id="get_configuration">}\href{../../rhdx/html/HDXObject.html#method-get_configuration}{\code{rhdx::HDXObject$get_configuration()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rhdx" data-topic="HDXObject" data-id="get_required_fields">}\href{../../rhdx/html/HDXObject.html#method-get_required_fields}{\code{rhdx::HDXObject$get_required_fields()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rhdx" data-topic="HDXObject" data-id="update_from_json">}\href{../../rhdx/html/HDXObject.html#method-update_from_json}{\code{rhdx::HDXObject$update_from_json()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rhdx" data-topic="HDXObject" data-id="update_from_yaml">}\href{../../rhdx/html/HDXObject.html#method-update_from_yaml}{\code{rhdx::HDXObject$update_from_yaml()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new Tag object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Vocabulary$new(initial_data = NULL, configuration = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{initial_data}}{list data with required fields to create a tag object}

\item{\code{configuration}}{Configuration configuration to use}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a Tag object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-as_list"></a>}}
\if{latex}{\out{\hypertarget{method-as_list}{}}}
\subsection{Method \code{as_list()}}{
Tag object to list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Vocabulary$as_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a list with of tag fields element
List of accepted tags on HDX
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-authorized_tags_name"></a>}}
\if{latex}{\out{\hypertarget{method-authorized_tags_name}{}}}
\subsection{Method \code{authorized_tags_name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Vocabulary$authorized_tags_name()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a vector of tags name
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print a Vocabulary object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Vocabulary$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Vocabulary$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
