% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration.R
\docType{data}
\name{Configuration}
\alias{Configuration}
\title{Configuration for HDX}
\description{
HDX Configuration allow to connect to an HDX server and setup project where you can interact with the HDX platform
}
\details{
**Methods**
  \describe{
    \item{`create(hdx_site, hdx_key, read_only, hdx_key_file, hdx_config, hdx_config_yaml, hdx_config_json, project_config, project_config_yaml, project_config_json, configuration)`}{
      Create a HDX Configuration 
    }
    \item{`setup(hdx_site, hdx_key, read_only, hdx_key_file, hdx_config, hdx_config_yaml, hdx_config_json, project_config, project_config_yaml, project_config_json, configuration)`}{
      Setup a HDX Configuration object
    }
    \item{`read()`}{
      Read a Configuration object
    }
    \item{`delete()`}{
      Delete configuration
    }
  }
}
\examples{
\dontrun{
set_rhdx_config(hdx_site = "demo")
get_rhd_config()
}

}
\keyword{datasets}
