% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdx_config.R
\docType{data}
\name{Configuration}
\alias{Configuration}
\title{Create an R6 reference object generator}
\description{
R6 objects are essentially environments, structured in a way that makes them
look like an object in a more typical object-oriented language than R. They
support public and private members, as well as inheritance across different
packages.
}
\details{
**Methods**
  \describe{
    \item{`create(path, query, disk, stream, ...)`}{
      Make a GET request
    }
    \item{`read(path, query, body, disk, stream, ...)`}{
      Make a POST request
    }
    \item{`delete(path, query, body, disk, stream, ...)`}{
      Make a PUT request
    }
    \item{`setup(path, query, body, disk, stream, ...)`}{
      Make a PATCH request
    }
    \item{`delete(path, query, body, disk, stream, ...)`}{
      Make a DELETE request
    }
    \item{`head(path, query, ...)`}{
      Make a HEAD request
    }
  }

Possible parameters (not all are allowed in each HTTP verb):
\itemize{
 \item create - URL path, appended to the base URL
 \item read - query terms, as a named list
 \item delete - body as an R list
 \item setup - one of form, multipart, json, or raw
}
}
\examples{
# A queue ---------------------------------------------------------
Configuration$create(hdx_site = "demo"),
conf <- Configuration$read()
conf

}
\keyword{datasets}
