% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{.count_datasets}
\alias{.count_datasets}
\alias{count_datasets}
\title{Count all datasets on HDX}
\usage{
.count_datasets(configuration = NULL)

count_datasets(configuration = NULL)
}
\arguments{
\item{configuration}{an HDX Configuration object}
}
\value{
An integer, the number of datasets
}
\description{
Count all datasets on HDX
}
