% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{create.Resource}
\alias{create.Resource}
\title{Create resource in HDX}
\usage{
create.Resource(resource, verbose = FALSE)
}
\arguments{
\item{resource}{Resource}
}
\value{
an HDX resource
}
\description{
Create resource in HDX
}
\examples{
\dontrun{
}
}
