% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{Dataset}
\alias{Dataset}
\title{HDX Dataset}
\description{
Dataset class containing all logic for accessing, creating, and updating datasets and associated resources.
}
\details{
Possible parameters (not all are allowed in each HTTP verb):
}
\examples{
# ---------------------------------------------------------
\dontrun{
set_rhdx_config(hdx_site = "prod")
acled_mali <- read_dataset("acled-conflict-data-for-africa-realtime-2016")
acled_mali

search_datasets("3W Somalia", rows = 5)

}

}
\keyword{datasets}
