% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{Dataset}
\alias{Dataset}
\title{HDX Dataset}
\description{
HDX Dataset
Dataset class containing all logic for accessing,
creating, and updating datasets and associated resources.
}
\section{Details}{

\strong{Methods}
\describe{
\item{\code{Dataset$new(initial_data, configuration)}}{
Dataset class enabling operations on datasets and associated resources
}

\item{\code{check_required_fields()}}{
Check that metadata for dataset and its resources is complete.
}

\item{\code{create(upload_resources)}}{
Check if dataset exists in HDX and if so, update it, otherwise create it
}

\item{\code{get_configuration(identifier, configuration)}}{
Returns the actual config used to get the dataset
}

\item{\code{get_dataset_date()}}{
Get dataset date as string.
}\preformatted{\\item\{`pull(identifier, configuration)`\}\{
  Reads the dataset given by identifier from HDX and returns Dataset object
\}

\\item\{`get_resource(index)`\}\{
  Get one resource from dataset by index
\}

\\item\{`get_resources()`\}\{
  Get dataset’s resources
\}

\\item\{`get_update_frequency()`\}\{
  Get expected update frequency.
\}
}

}
}

\examples{
\dontrun{
 set_rhdx_config(hdx_site = "prod")
 acled_mali_rs <- pull_dataset("acled-data-for-mali")
 acled_mali_rs
}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Dataset$new()}}
\item \href{#method-get_resource}{\code{Dataset$get_resource()}}
\item \href{#method-get_resources}{\code{Dataset$get_resources()}}
\item \href{#method-delete_resource}{\code{Dataset$delete_resource()}}
\item \href{#method-delete_resources}{\code{Dataset$delete_resources()}}
\item \href{#method-browse}{\code{Dataset$browse()}}
\item \href{#method-get_configuration}{\code{Dataset$get_configuration()}}
\item \href{#method-get_dataset_date}{\code{Dataset$get_dataset_date()}}
\item \href{#method-set_dataset_date}{\code{Dataset$set_dataset_date()}}
\item \href{#method-get_update_frequency}{\code{Dataset$get_update_frequency()}}
\item \href{#method-get_tags}{\code{Dataset$get_tags()}}
\item \href{#method-get_locations}{\code{Dataset$get_locations()}}
\item \href{#method-get_maintainer}{\code{Dataset$get_maintainer()}}
\item \href{#method-get_organization}{\code{Dataset$get_organization()}}
\item \href{#method-get_showcases}{\code{Dataset$get_showcases()}}
\item \href{#method-set_organization}{\code{Dataset$set_organization()}}
\item \href{#method-is_requestable}{\code{Dataset$is_requestable()}}
\item \href{#method-get_required_fields}{\code{Dataset$get_required_fields()}}
\item \href{#method-check_required_fields}{\code{Dataset$check_required_fields()}}
\item \href{#method-as_list}{\code{Dataset$as_list()}}
\item \href{#method-print}{\code{Dataset$print()}}
\item \href{#method-clone}{\code{Dataset$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$new(initial_data = NULL, configuration = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_resource"></a>}}
\subsection{Method \code{get_resource()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$get_resource(index)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_resources"></a>}}
\subsection{Method \code{get_resources()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$get_resources()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_resource"></a>}}
\subsection{Method \code{delete_resource()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$delete_resource(index = 1L)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_resources"></a>}}
\subsection{Method \code{delete_resources()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$delete_resources()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-browse"></a>}}
\subsection{Method \code{browse()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$browse()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_configuration"></a>}}
\subsection{Method \code{get_configuration()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$get_configuration()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_dataset_date"></a>}}
\subsection{Method \code{get_dataset_date()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$get_dataset_date()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_dataset_date"></a>}}
\subsection{Method \code{set_dataset_date()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$set_dataset_date(date, format = "\%m/\%d/\%Y")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_update_frequency"></a>}}
\subsection{Method \code{get_update_frequency()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$get_update_frequency()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_tags"></a>}}
\subsection{Method \code{get_tags()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$get_tags()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_locations"></a>}}
\subsection{Method \code{get_locations()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$get_locations()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_maintainer"></a>}}
\subsection{Method \code{get_maintainer()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$get_maintainer()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_organization"></a>}}
\subsection{Method \code{get_organization()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$get_organization()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_showcases"></a>}}
\subsection{Method \code{get_showcases()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$get_showcases()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_organization"></a>}}
\subsection{Method \code{set_organization()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$set_organization(organization)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_requestable"></a>}}
\subsection{Method \code{is_requestable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$is_requestable()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_required_fields"></a>}}
\subsection{Method \code{get_required_fields()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$get_required_fields()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-check_required_fields"></a>}}
\subsection{Method \code{check_required_fields()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$check_required_fields()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-as_list"></a>}}
\subsection{Method \code{as_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$as_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
