% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{Dataset}
\alias{Dataset}
\title{HDX Dataset}
\description{
Dataset class containing all logic for accessing,
creating, and updating datasets and associated resources.
}
\details{
**Methods**
  \describe{
    \item{`Dataset$new(initial_data, configuration)`}{
      Dataset class enabling operations on datasets and associated resources
    }

   \item{`add_resource(resource, ignore_datasetid = FALSE)`}{
      Add new resource in dataset with new metadata
    }

  \item{`check_required_fields()`}{
     Check that metadata for dataset and its resources is complete.
    }

   \item{`create_in_hdx(upload_resources)`}{
    Check if dataset exists in HDX and if so, update it, otherwise create it
    }

   \item{`delete_from_hdx(identifier, configuration)`}{
      Reads the dataset given by identifier
      from HDX and returns Dataset object
    }

   \item{`get_configuration(identifier, configuration)`}{
      Returns the actual config used to get the dataset
    }

   \item{`get_dataset_date()`}{
      Get dataset date as string.
    }

    \item{`read_from_hdx(identifier, configuration)`}{
      Reads the dataset given by identifier from HDX and returns Dataset object
    }

    \item{`get_resource(index)`}{
      Get one resource from dataset by index
    }

    \item{`get_resources()`}{
      Get dataset’s resources
    }

    \item{`get_update_frequency()`}{
      Get expected update frequency.
    }

    \item{`set_update_frequency()`}{
      Set expected update frequency.
    }


  }
}
\section{Details}{

}

\examples{
\dontrun{
 set_rhdx_config(hdx_site = "prod")
 acled_mali_rs <- pull_dataset("acled-data-for-mali")
 acled_mali_rs
}

}
\keyword{datasets}
