% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{delete_resource}
\alias{delete_resource}
\title{Delete resource from dataset}
\usage{
delete_resource(dataset, index)
}
\arguments{
\item{dataset}{}
}
\value{
A Dataset
}
\description{
Delete resource from dataset
}
\details{
Delete resource from dataset
}
\examples{
\dontrun{
 # Setting the config to use HDX default server
 delete_resource(dataset, 1) # remove first resource
}

}
