% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{delete_resource}
\alias{delete_resource}
\title{Delete delete resource}
\usage{
delete_resource(dataset, index)
}
\arguments{
\item{dataset}{}
}
\value{
A list of HDX datasets
}
\description{
Sets the configuration settings for using rhdx.
}
\details{
Delete resource from dataset
}
\examples{

\dontrun{
 # Setting the config to use HDX default server
 delete_resource(dataset, 1) # first resource
}

}
