% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{.list_datasets}
\alias{.list_datasets}
\alias{list_datasets}
\title{List datasets}
\usage{
.list_datasets(limit = NULL, offset = NULL, configuration = NULL)

list_datasets(limit = NULL, offset = NULL, configuration = NULL)
}
\arguments{
\item{limit}{integer; limit}

\item{offset}{integer; offset}

\item{configuration}{a Configuration object}
}
\value{
A vector of datasets names
}
\description{
List datasets
}
\examples{
\dontrun{
# Setting the config to use HDX default server
 set_rhdx_config()
 list_datasets(limit = 10L)
}
}
