% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{list_users}
\alias{list_users}
\title{List all users}
\usage{
list_users(sort = "name asc", all_fields = FALSE,
  configuration = NULL, ...)
}
\arguments{
\item{sort}{Logical user sorted is TRUE}

\item{all_fields}{Logical if TRUE get all field}

\item{configuration}{Configuration the configuration to use}

\item{...}{Extra parameters}
}
\description{
List all users
}
