% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{.list_users}
\alias{.list_users}
\alias{list_users}
\title{List all users}
\usage{
.list_users(order_by = "number_created_packages", configuration = NULL,
  ...)

list_users(order_by = "number_created_packages", configuration = NULL,
  ...)
}
\arguments{
\item{configuration}{Configuration the configuration to use}

\item{...}{Extra parameters}

\item{sort}{Logical user sorted is TRUE}

\item{all_fields}{Logical if TRUE get all field}
}
\description{
List all users
}
