% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location.R
\docType{data}
\name{Location}
\alias{Location}
\title{HDX Location}
\description{
HDX location mostly countries
}
\details{
\strong{Methods}
\describe{
}
}
\examples{
# ---------------------------------------------------------
\dontrun{
}
}
\keyword{datasets}
