% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location.R
\name{Location}
\alias{Location}
\title{HDX Location}
\description{
HDX location mostly countries
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Location$new()}}
\item \href{#method-get_required_fields}{\code{Location$get_required_fields()}}
\item \href{#method-check_required_fields}{\code{Location$check_required_fields()}}
\item \href{#method-browse}{\code{Location$browse()}}
\item \href{#method-as_list}{\code{Location$as_list()}}
\item \href{#method-print}{\code{Location$print()}}
\item \href{#method-clone}{\code{Location$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Location$new(initial_data = NULL, configuration = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_required_fields"></a>}}
\subsection{Method \code{get_required_fields()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Location$get_required_fields()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-check_required_fields"></a>}}
\subsection{Method \code{check_required_fields()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Location$check_required_fields()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-browse"></a>}}
\subsection{Method \code{browse()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Location$browse()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-as_list"></a>}}
\subsection{Method \code{as_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Location$as_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Location$print(x, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Location$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
