% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location.R
\docType{data}
\name{Location}
\alias{Location}
\title{Create and manipulate HDX Location}
\description{
HDX location mostly countries
}
\details{
**Methods**
  \describe{
  }
}
\examples{
# ---------------------------------------------------------
\dontrun{
Configuration$create(hdx_site = "demo")
org <- Location$read_from_hdx("ocha-mali", include_dataset = TRUE)
org
}

}
\keyword{datasets}
