% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organization.R
\docType{data}
\name{Organization}
\alias{Organization}
\title{Create and manipulate HDX Organization}
\description{
Create and manipulate HDX Organization
}
\details{
**Methods**
  \describe{
  }
}
\examples{
\dontrun{
}

}
\keyword{datasets}
