% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{.pull_dataset}
\alias{.pull_dataset}
\alias{pull_dataset}
\title{Pull HDX dataset into R}
\usage{
.pull_dataset(identifier, configuration = NULL, ...)

pull_dataset(identifier, configuration = NULL, ...)
}
\arguments{
\item{identifier}{Character dataset keyword}

\item{configuration}{a Configuration object}

\item{...}{Extra parameters}
}
\value{
Dataset the dataset
}
\description{
Read an HDX dataset from its name or id
}
\examples{
\dontrun{
# Setting the config to use HDX default server
 set_rhdx_config()
 res <- pull_dataset("mali-3wop")
 res
}
}
