% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location.R
\name{pull_location}
\alias{pull_location}
\title{Read an HDX location}
\usage{
pull_location(identifier = NULL, include_datasets = FALSE,
  configuration = NULL, ...)
}
\arguments{
\item{identifier}{Character location uuid}

\item{configuration}{Configuration a configuration object}

\item{...}{Extra parameters}
}
\value{
Location
}
\description{
Read an HDX location
}
\examples{
\dontrun{
#Setting the config to use HDX default server
 set_rhdx_config()
 res <- pull_location("mli")
 res
}
}
