% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{.pull_resource}
\alias{.pull_resource}
\alias{pull_resource}
\title{Read an HDX resource}
\usage{
.pull_resource(identifier = NULL, configuration = NULL, ...)

pull_resource(identifier = NULL, configuration = NULL, ...)
}
\arguments{
\item{identifier}{character resource uuid}

\item{configuration}{an HDX configuration object}

\item{...}{Extra parameters}
}
\value{
Resource
}
\description{
Read an HDX resource
}
\examples{
\dontrun{
#Setting the config to use HDX default server
 set_rhdx_config()
 res <- read_resource("98aa1742-b5d3-40c3-94c6-01e31ded6e84")
 res
}
}
