% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{pull_user}
\alias{pull_user}
\title{Read an HDX user}
\usage{
pull_user(identifier = NULL, include_datasets = FALSE,
  configuration = NULL, ...)
}
\arguments{
\item{identifier}{character user keyword}

\item{include_datasets}{Logical, if TRUE add datasets}

\item{configuration}{a Configuration object}

\item{...}{Extra parameters}
}
\value{
User the user
}
\description{
Read an HDX user from its name or id
}
\examples{
\dontrun{
# Setting the config to use HDX default server
 set_rhdx_config()
 res <- pull_user("xxxx")
 res
}
}
