% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{push_dataset}
\alias{push_dataset}
\title{Create dataset in HDX}
\usage{
push_dataset(dataset, upload_resources = TRUE, verbose = FALSE)
}
\arguments{
\item{dataset}{Dataset the dataset we want to push online}

\item{upload_resources}{Upload resources online}

\item{verbose}{Logical Silent output if FALSE}
}
\value{
an HDX dataset
}
\description{
Create dataset in HDX
}
