% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{Resource}
\alias{Resource}
\title{HDX Resource}
\description{
HDX Resource
}
\examples{
\dontrun{
  set_rhdx_config()
  resource <- pull_resource("98aa1742-b5d3-40c3-94c6-01e31ded6e84")
  resource
}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Resource$new()}}
\item \href{#method-download}{\code{Resource$download()}}
\item \href{#method-download_folder}{\code{Resource$download_folder()}}
\item \href{#method-read_resource}{\code{Resource$read_resource()}}
\item \href{#method-get_layers}{\code{Resource$get_layers()}}
\item \href{#method-get_sheets}{\code{Resource$get_sheets()}}
\item \href{#method-get_dataset}{\code{Resource$get_dataset()}}
\item \href{#method-get_file_to_upload}{\code{Resource$get_file_to_upload()}}
\item \href{#method-get_required_fields}{\code{Resource$get_required_fields()}}
\item \href{#method-check_required_field}{\code{Resource$check_required_field()}}
\item \href{#method-get_file_type}{\code{Resource$get_file_type()}}
\item \href{#method-get_format}{\code{Resource$get_format()}}
\item \href{#method-as_list}{\code{Resource$as_list()}}
\item \href{#method-browse}{\code{Resource$browse()}}
\item \href{#method-print}{\code{Resource$print()}}
\item \href{#method-clone}{\code{Resource$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Resource$new(initial_data = NULL, configuration = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-download"></a>}}
\subsection{Method \code{download()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Resource$download(
  folder = NULL,
  filename = NULL,
  quiet = TRUE,
  force = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-download_folder"></a>}}
\subsection{Method \code{download_folder()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Resource$download_folder()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read_resource"></a>}}
\subsection{Method \code{read_resource()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Resource$read_resource(
  sheet = NULL,
  layer = NULL,
  folder = NULL,
  simplify_json = TRUE,
  force_download = FALSE,
  quiet = TRUE,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_layers"></a>}}
\subsection{Method \code{get_layers()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Resource$get_layers(folder = NULL, quiet = TRUE, force_download = FALSE, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_sheets"></a>}}
\subsection{Method \code{get_sheets()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Resource$get_sheets(folder = NULL, quiet = TRUE, force_download = FALSE, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_dataset"></a>}}
\subsection{Method \code{get_dataset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Resource$get_dataset()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_file_to_upload"></a>}}
\subsection{Method \code{get_file_to_upload()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Resource$get_file_to_upload()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_required_fields"></a>}}
\subsection{Method \code{get_required_fields()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Resource$get_required_fields()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-check_required_field"></a>}}
\subsection{Method \code{check_required_field()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Resource$check_required_field(check_dataset_id = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_file_type"></a>}}
\subsection{Method \code{get_file_type()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Resource$get_file_type()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_format"></a>}}
\subsection{Method \code{get_format()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Resource$get_format()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-as_list"></a>}}
\subsection{Method \code{as_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Resource$as_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-browse"></a>}}
\subsection{Method \code{browse()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Resource$browse()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Resource$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Resource$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
