% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\docType{data}
\name{Resource}
\alias{Resource}
\title{HDX Resource}
\description{
HDX Resource
}
\details{
\strong{Methods}
\describe{
}
}
\examples{
\dontrun{
  set_rhdx_config()
  resource <- read_resource("98aa1742-b5d3-40c3-94c6-01e31ded6e84")
  resource
}
}
\keyword{datasets}
