% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{Resource}
\alias{Resource}
\title{HDX Resource}
\description{
HDX Resource

HDX Resource
}
\details{
HDX Resource, it contains all the logic for creating, checking,
and updating resources
}
\section{Super class}{
\code{\link[rhdx:HDXObject]{rhdx::HDXObject}} -> \code{Resource}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{placeholder for Resource field element}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Resource$new()}}
\item \href{#method-download}{\code{Resource$download()}}
\item \href{#method-download_folder}{\code{Resource$download_folder()}}
\item \href{#method-read_resource}{\code{Resource$read_resource()}}
\item \href{#method-get_layers}{\code{Resource$get_layers()}}
\item \href{#method-get_sheets}{\code{Resource$get_sheets()}}
\item \href{#method-get_dataset}{\code{Resource$get_dataset()}}
\item \href{#method-get_required_fields}{\code{Resource$get_required_fields()}}
\item \href{#method-check_required_field}{\code{Resource$check_required_field()}}
\item \href{#method-get_format}{\code{Resource$get_format()}}
\item \href{#method-as_list}{\code{Resource$as_list()}}
\item \href{#method-browse}{\code{Resource$browse()}}
\item \href{#method-print}{\code{Resource$print()}}
\item \href{#method-clone}{\code{Resource$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="rhdx" data-topic="HDXObject" data-id="get_configuration">}\href{../../rhdx/html/HDXObject.html#method-get_configuration}{\code{rhdx::HDXObject$get_configuration()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rhdx" data-topic="HDXObject" data-id="update_from_json">}\href{../../rhdx/html/HDXObject.html#method-update_from_json}{\code{rhdx::HDXObject$update_from_json()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rhdx" data-topic="HDXObject" data-id="update_from_yaml">}\href{../../rhdx/html/HDXObject.html#method-update_from_yaml}{\code{rhdx::HDXObject$update_from_yaml()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new Resource object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Resource$new(initial_data = NULL, configuration = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{initial_data}}{list with required field to create a resource}

\item{\code{configuration}}{a Configuration object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new Resource object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-download"></a>}}
\if{latex}{\out{\hypertarget{method-download}{}}}
\subsection{Method \code{download()}}{
Download a HDX resource
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Resource$download(
  folder = NULL,
  filename = NULL,
  quiet = TRUE,
  force = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{folder}}{a character, folder to save the dataset}

\item{\code{filename}}{a character, filename of the dataset}

\item{\code{quiet}}{a logical value, silent download if TRUE}

\item{\code{force}}{a logical value, force download}

\item{\code{...}}{other \code{download.file} parameters}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a character, the file path
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-download_folder"></a>}}
\if{latex}{\out{\hypertarget{method-download_folder}{}}}
\subsection{Method \code{download_folder()}}{
Get the download folder for the latest downloaded resource
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Resource$download_folder()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a character, folder with the latest downloaded resource
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read_resource"></a>}}
\if{latex}{\out{\hypertarget{method-read_resource}{}}}
\subsection{Method \code{read_resource()}}{
Read a Resource object directly into memory
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Resource$read_resource(
  sheet = NULL,
  layer = NULL,
  format = NULL,
  download_folder = NULL,
  simplify_json = TRUE,
  force_download = FALSE,
  quiet_download = TRUE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sheet}}{a character value, only for resource in Excel format}

\item{\code{layer}}{a character value, only for spatial (vector) resource}

\item{\code{format}}{a character value, file format;}

\item{\code{download_folder}}{a character value, folder to save the downloaded resource}

\item{\code{simplify_json}}{a logical value}

\item{\code{force_download}}{a logical value, if TRUE force download}

\item{\code{quiet_download}}{a logical value, if TRUE silent download}

\item{\code{...}}{other parameters}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a \code{tibble}, a \code{sf}, a \code{stars} or a \code{list} depending
on the type of resource read
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_layers"></a>}}
\if{latex}{\out{\hypertarget{method-get_layers}{}}}
\subsection{Method \code{get_layers()}}{
Get spatial (vector) resource list of layers
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Resource$get_layers(
  format = NULL,
  download_folder = NULL,
  quiet_download = TRUE,
  force_download = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{format}}{character; file format}

\item{\code{download_folder}}{a character value, folder to save the downloaded resource}

\item{\code{quiet_download}}{a logical value, if TRUE silent download}

\item{\code{force_download}}{a logical value, if TRUE force download}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a the list of layers available in the resource
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_sheets"></a>}}
\if{latex}{\out{\hypertarget{method-get_sheets}{}}}
\subsection{Method \code{get_sheets()}}{
Get the list of sheets name of resource
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Resource$get_sheets(
  format = NULL,
  download_folder = NULL,
  quiet_download = TRUE,
  force_download = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{format}}{character; file format}

\item{\code{download_folder}}{a character value, folder to save the downloaded resource}

\item{\code{quiet_download}}{a logical value, if TRUE silent download}

\item{\code{force_download}}{a logical value, if TRUE force download}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a the list of layers available in the resource
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_dataset"></a>}}
\if{latex}{\out{\hypertarget{method-get_dataset}{}}}
\subsection{Method \code{get_dataset()}}{
Get the resource dataset.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Resource$get_dataset()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a Dataset, the dataset containing the resource
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_required_fields"></a>}}
\if{latex}{\out{\hypertarget{method-get_required_fields}{}}}
\subsection{Method \code{get_required_fields()}}{
Get dataset required fields
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Resource$get_required_fields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list of required fields for a resource
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-check_required_field"></a>}}
\if{latex}{\out{\hypertarget{method-check_required_field}{}}}
\subsection{Method \code{check_required_field()}}{
Check dataset required field
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Resource$check_required_field(check_dataset_id = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{check_dataset_id}}{logical whether to check or not dataset id}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a logical value, TRUE if the the resource is not missing
a required field and throws an error otherwise
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_format"></a>}}
\if{latex}{\out{\hypertarget{method-get_format}{}}}
\subsection{Method \code{get_format()}}{
Get the file format
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Resource$get_format()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a character, the file format of the resource
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-as_list"></a>}}
\if{latex}{\out{\hypertarget{method-as_list}{}}}
\subsection{Method \code{as_list()}}{
Get resource field into list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Resource$as_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a list with resource field
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-browse"></a>}}
\if{latex}{\out{\hypertarget{method-browse}{}}}
\subsection{Method \code{browse()}}{
Browse the resource page on HDX
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Resource$browse()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print a Resource object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Resource$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Resource$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
