% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching.R
\name{rhdx_cache}
\alias{rhdx_cache}
\title{Caching}
\description{
Manage cached HDX Dataset with \pkg{hoardr}
}
\details{
The default cache directory is
\code{~/.cache/R/rhdx_cache}, but you can set
your own path using \code{cache_path_set()}

\code{cache_delete} only accepts 1 file name, while
\code{cache_delete_all} doesn't accept any names, but deletes all files.
For deleting many specific files, use \code{cache_delete} in a \code{\link[=lapply]{lapply()}}
type call
}
\section{Useful user functions}{

\itemize{
\item \code{rhdx_cache$cache_path_get()} get cache path
\item \code{rhdx_cache$cache_path_set()} set cache path
\item \code{rhdx_cache$list()} returns a character vector of full path file names
\item \code{rhdx_cache$files()} returns file objects with metadata
\item \code{rhdx_cache$details()} returns files with details
\item \code{rhdx_cache$delete()} delete specific files
\item \code{rhdx_cache$delete_all()} delete all files, returns nothing
}
}

\examples{
\dontrun{
rhdx_cache

# list files in cache
rhdx_cache$list()

# delete certain database files
# rhdx_cache$delete("file path")
# rhdx_cache$list()

# delete all files in cache
# rhdx_cache$delete_all()
# rhdx_cache$list()

# set a different cache path from the default
}
}
