% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching.R
\name{rhdx_cache_delete}
\alias{rhdx_cache_delete}
\title{Delete file from cache}
\usage{
rhdx_cache_delete(file)
}
\arguments{
\item{file}{Character, the file to delete}
}
\description{
Delete file from cache
}
