% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{search_datasets}
\alias{search_datasets}
\title{Search datasets on HDX}
\usage{
search_datasets(query = "*:*", rows = 10L, page_size = 1000L,
  configuration = NULL, ...)
}
\arguments{
\item{query}{Character Query (in Solr format). Defaults to ‘*:*’}

\item{rows}{Number of matching rows to return. Defaults to 10.}

\item{page_size}{Integer Size of page to return. Defaults to 1000.}

\item{configuration}{Configuration object.}
}
\value{
A list of HDX datasets
}
\description{
Sets the configuration settings for using rhdx.
}
\details{
Allow to search and find on HDX
}
\examples{

\dontrun{
 # Setting the config to use HDX default server
 search_datasets("displaced nigeria", rows = 3L)
}

}
