% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{User}
\alias{User}
\title{HDX User}
\description{
HDX user
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{User$new()}}
\item \href{#method-update_from_yaml}{\code{User$update_from_yaml()}}
\item \href{#method-update_from_json}{\code{User$update_from_json()}}
\item \href{#method-as_list}{\code{User$as_list()}}
\item \href{#method-print}{\code{User$print()}}
\item \href{#method-clone}{\code{User$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$new(initial_data = NULL, configuration = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_from_yaml"></a>}}
\subsection{Method \code{update_from_yaml()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$update_from_yaml(hdx_user_static_yaml)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_from_json"></a>}}
\subsection{Method \code{update_from_json()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$update_from_json(hdx_user_static_json)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-as_list"></a>}}
\subsection{Method \code{as_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$as_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
