% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\docType{data}
\name{User}
\alias{User}
\title{HDX User}
\description{
HDX user
}
\details{
**Methods**
  \describe{
  }
}
\examples{
Configuration$create(hdx_site = "demo")
User$list_all_users()

}
\keyword{datasets}
