% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{User}
\alias{User}
\title{HDX User}
\description{
HDX User

HDX User
}
\details{
HDX user
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{placeholder for Dataset field element}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{User$new()}}
\item \href{#method-as_list}{\code{User$as_list()}}
\item \href{#method-print}{\code{User$print()}}
\item \href{#method-clone}{\code{User$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$new(initial_data = NULL, configuration = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{initial_data}}{list of field required to create a dataset}

\item{\code{configuration}}{Configuration configuration to use}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a new User object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-as_list"></a>}}
\if{latex}{\out{\hypertarget{method-as_list}{}}}
\subsection{Method \code{as_list()}}{
Get dataset field into list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$as_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a list with dataset field
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print a User object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
